Use float32 precision for python test

+++ algorithms/averageMimuData/_tests/test_averageMimuDataF32.py
@@ -31,9 +31,9 @@ def test_average_mimu_data():
     module.modelTag = "averageMimuData"
 
     # Define parameters
-    dcm_pltf_to_body = RigidBodyKinematics.euler3212C([0.01, -0.04, 0.06])
+    dcm_pltf_to_body = RigidBodyKinematics.euler3212C([0.01, -0.04, 0.06]).astype(np.float32)
     module.setDcmPltfToBdy(dcm_pltf_to_body)
-    duration_window = 1.0e10
+    duration_window = 1.0e10 # make the time window so huge that average is taken over all packets
     module.setAveragingWindow(duration_window)
 
     # Initialize message for message connection
@@ -58,18 +58,18 @@ def test_average_mimu_data():
     meas_time = 0
     sim_time = 0
     counter = 0
-    gyro_sum = np.zeros(3)
-    accel_sum = np.zeros(3)
-    average_imu_output = np.zeros([num_fsw_steps, 3])
-    average_acc_output = np.zeros([num_fsw_steps, 3])
+    gyro_sum = np.zeros(3, dtype=np.float32)
+    accel_sum = np.zeros(3,dtype=np.float32)
+    average_imu_output = np.zeros([num_fsw_steps, 3], dtype=np.float32)
+    average_acc_output = np.zeros([num_fsw_steps, 3], dtype=np.float32)
     for index in range(num_fsw_steps):
         for _ in range(samples_per_fsw):
             meas_time += delta_time
             acc_data.accPkts[counter].measTime = macros.sec2nano(meas_time)
-            random_array = np.random.normal(loc=2, scale=1, size=3)
+            random_array = np.random.normal(loc=2, scale=1, size=3).astype(np.float32)
             gyro_sum += random_array
             acc_data.accPkts[counter].gyro_B = random_array.tolist()
-            random_array = np.random.normal(loc=2, scale=1, size=3)
+            random_array = np.random.normal(loc=2, scale=1, size=3).astype(np.float32)
             accel_sum += random_array
             acc_data.accPkts[counter].accel_B = random_array.tolist()
             counter += 1
